+++ tests/tools/app_store_connect/test_from_cli_args.py
@@ -2,7 +2,6 @@ from __future__ import annotations
 
 import argparse
 import os
-import pathlib
 from tempfile import NamedTemporaryFile
 from unittest import mock
 
@@ -11,32 +10,6 @@ import pytest
 from codemagic.tools.app_store_connect import AppStoreConnect
 from codemagic.tools.app_store_connect import AppStoreConnectArgument
 from codemagic.tools.app_store_connect import Types
-
-
-@pytest.fixture(autouse=True)
-def register_args(cli_argument_group):
-    for arg in AppStoreConnect.CLASS_ARGUMENTS:
-        arg.register(cli_argument_group)
-
-
-@pytest.fixture()
-def namespace_kwargs():
-    ns_kwargs = {
-        'action': 'list-devices',
-        AppStoreConnectArgument.CERTIFICATES_DIRECTORY.key:
-            AppStoreConnectArgument.CERTIFICATES_DIRECTORY.get_default(),
-        AppStoreConnectArgument.PROFILES_DIRECTORY.key: AppStoreConnectArgument.PROFILES_DIRECTORY.get_default(),
-        AppStoreConnectArgument.LOG_REQUESTS.key: True,
-        AppStoreConnectArgument.JSON_OUTPUT.key: False,
-        AppStoreConnectArgument.ISSUER_ID.key: Types.IssuerIdArgument('issuer-id'),
-        AppStoreConnectArgument.KEY_IDENTIFIER.key: Types.KeyIdentifierArgument('key-identifier'),
-        AppStoreConnectArgument.PRIVATE_KEY.key: Types.PrivateKeyArgument('-----BEGIN PRIVATE KEY-----'),
-    }
-    for arg in AppStoreConnect.CLASS_ARGUMENTS:
-        if not hasattr(arg.type, 'environment_variable_key'):
-            continue
-        os.environ.pop(arg.type.environment_variable_key, None)
-    return ns_kwargs
 
 
 def _test_missing_argument(argument, _namespace_kwargs):
@@ -145,55 +118,3 @@ def _do_private_key_assertions(private_key_value, moc_appstore_api_client, cli_n
     _, _, private_key_arg = moc_appstore_api_client.call_args[0]
     assert isinstance(private_key_arg, str)
     assert private_key_arg == private_key_value
-
-
-def test_publish_action_without_app_store_connect_key(namespace_kwargs):
-    namespace_kwargs.update({
-        AppStoreConnectArgument.ISSUER_ID.key: None,
-        AppStoreConnectArgument.KEY_IDENTIFIER.key: None,
-        AppStoreConnectArgument.PRIVATE_KEY.key: None,
-        'action': 'publish',
-    })
-    cli_args = argparse.Namespace(**namespace_kwargs)
-    _ = AppStoreConnect.from_cli_args(cli_args)
-
-
-@pytest.mark.parametrize('missing_argument', [
-    AppStoreConnectArgument.ISSUER_ID,
-    AppStoreConnectArgument.KEY_IDENTIFIER,
-    AppStoreConnectArgument.PRIVATE_KEY,
-])
-def test_publish_action_without_app_store_connect_key_testflight_submit(missing_argument, namespace_kwargs):
-    namespace_kwargs.update({missing_argument.key: None, 'action': 'publish'})
-    cli_args = argparse.Namespace(**namespace_kwargs)
-    app_store_connect = AppStoreConnect.from_cli_args(cli_args)
-    with pytest.raises(argparse.ArgumentError) as error_info:
-        app_store_connect.publish(
-            application_package_path_patterns=[pathlib.Path('path.pattern')],
-            submit_to_testflight=True,
-        )
-    assert missing_argument.flag in error_info.value.argument_name
-
-
-@mock.patch('codemagic.tools._app_store_connect.actions.publish_action.Altool')
-def test_publish_action_with_username_and_password(_mock_altool, namespace_kwargs):
-    namespace_kwargs.update({
-        AppStoreConnectArgument.ISSUER_ID.key: None,
-        AppStoreConnectArgument.KEY_IDENTIFIER.key: None,
-        AppStoreConnectArgument.PRIVATE_KEY.key: None,
-        'action': 'publish',
-    })
-
-    cli_args = argparse.Namespace(**namespace_kwargs)
-    with mock.patch.object(AppStoreConnect, 'find_paths') as mock_find_paths, \
-         mock.patch.object(AppStoreConnect, '_get_publishing_application_packages') as mock_get_packages:
-        mock_get_packages.return_value = []
-        mock_find_paths.return_value = []
-
-        patterns = [pathlib.Path('path.pattern')]
-        AppStoreConnect.from_cli_args(cli_args).publish(
-            application_package_path_patterns=patterns,
-            apple_id='name@example.com',
-            app_specific_password=Types.AppSpecificPassword('xxxx-yyyy-zzzz-wwww'),
-        )
-        mock_get_packages.assert_called_with(patterns)